06/23/2023

+++ movieManiaDatabase/Movies.sql
@@ -2,6 +2,6 @@
 (
 	[movie_id] INT IDENTITY(1,1) NOT NULL PRIMARY KEY, 
     [title] VARCHAR(MAX) NULL, 
-    [price] FLOAT NULL, 
+    [price] DECIMAL NULL, 
     [stock] INT NULL
 )
